refactor:blog.dashboard part and change style

+++ src/app/api/posts/[id]/route.ts
@@ -50,6 +50,7 @@ export const GET = async (request: Request) => {
     if (!id) {
       return new NextResponse("ID is required", { status: 400 });
     }
+
     const post = await Post.findById(id);
     if (!post) {
       return new NextResponse("Post not found", { status: 404 });
@@ -62,18 +63,22 @@ export const GET = async (request: Request) => {
   }
 };
 
-
-
-export const DELETE = async (request: any, { params }: any) => {
-  const { id } = params;
+export const DELETE = async (request: Request) => {
+  const url = new URL(request.url);
+  const id = url.pathname.split("/").pop();
 
   try {
     await connect();
+
+    if (!id) {
+      return new NextResponse("ID is required", { status: 400 });
+    }
 
     await Post.findByIdAndDelete(id);
 
     return new NextResponse("Post has been deleted", { status: 200 });
   } catch (err) {
+    console.error("Database Error:", err);
     return new NextResponse("Database Error", { status: 500 });
   }
 };
